Crate panic_probe

Panic handler for probe-run.

When this panic handler is used, panics will make probe-run print a backtrace and exit with a non-zero status code, indicating failure. This building block can be used to run on-device tests.

Panic Messages

By default, panic-probe ignores the panic message. You can enable one of the following features to print it instead:

print-rtt: Prints the panic message over plain RTT (via rtt-target). RTT must be initialized by the app.

print-defmt: Prints the panic message via defmt’s transport (note that defmt will not be used to efficiently format the message).

Functions

hard_fault

Trigger a HardFault via udf instruction.